Move create_app back in cornac/__init__.py
