+++ service/README.md
@@ -25,8 +25,8 @@ The `CORNAC_SETTINGS` environment variable point to a python file containing
 regular
 [Flask configuration](http://flask.pocoo.org/docs/1.0/config/#configuring-from-files)
 and cornac configuration. [Default cornac
-configuration](cornac/flask/default_config.py) is commented. A
-[poc.cfg](poc.cfg) configuration file can be a good starting point.
+configuration](cornac/default_config.py) is commented. A [poc.cfg](poc.cfg)
+configuration file can be a good starting point.
 
 Further prerequisites depends on the infrastructure provider.
