+++ service/cornac/cli.py
@@ -36,7 +36,7 @@ class KnownError(Exception):
 
 def create_app():
     # Fake factory for Flask app.
-    from .flask import create_app as real_create_app
+    from . import create_app as real_create_app
     app = real_create_app()
     from .web import rds, fallback
     app.register_blueprint(rds)
